% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zinc_hanson_means}
\alias{zinc_hanson_means}
\title{Example founder means from ZINC Hanson experiment}
\format{
A tibble with 3,987,264 rows and 6 variables:
\describe{
\item{chr}{Chromosome}
\item{pos}{Genomic position (bp)}
\item{TRT}{Treatment (C or Z)}
\item{REP}{Replicate number}
\item{founder}{Founder strain}
\item{freq}{Allele frequency}
}
}
\usage{
zinc_hanson_means
}
\description{
This dataset contains founder mean frequencies for the ZINC Hanson experiment.
}
\keyword{datasets}
