% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zinc_hanson_pseudoscan}
\alias{zinc_hanson_pseudoscan}
\title{Example QTL scan results from ZINC Hanson experiment}
\format{
A tibble with 25,977 rows and 8 variables:
\describe{
\item{chr}{Chromosome}
\item{pos}{Genomic position (bp)}
\item{Wald_log10p}{-log10(p) from Wald test}
\item{Pseu_log10p}{-log10(p) from Pseudoscan}
\item{Falc_H2}{Falconer's H2}
\item{Cutl_H2}{Cutler's H2}
\item{avg.var}{Average variance}
\item{cM}{Genetic position (cM)}
}
}
\usage{
zinc_hanson_pseudoscan
}
\description{
This dataset contains QTL scan results for the ZINC Hanson experiment.
}
\keyword{datasets}
